moved view part of the login controller back to the login controller file
